Moved physics builder classes into a new package
