+++ kot/extensions/physics/builder/BodyBuilder.kt
@@ -1,9 +1,10 @@
-package kot.extensions.physics
+package kot.extensions.physics.builder
 
 import com.badlogic.gdx.physics.box2d.Body
 import com.badlogic.gdx.physics.box2d.BodyDef
 import com.badlogic.gdx.physics.box2d.Fixture
 import com.badlogic.gdx.physics.box2d.FixtureDef
+import kot.extensions.physics.PhysicsWorld
 import kotlin.reflect.jvm.internal.impl.load.kotlin.JvmType
 
 class BodyBuilder(private var world: PhysicsWorld) {
