+++ kot/extensions/physics/builder/FixtureDefBuilder.kt
@@ -1,10 +1,11 @@
-package kot.extensions.physics
+package kot.extensions.physics.builder
 
 import com.badlogic.gdx.math.Vector2
 import com.badlogic.gdx.physics.box2d.ChainShape
 import com.badlogic.gdx.physics.box2d.CircleShape
 import com.badlogic.gdx.physics.box2d.FixtureDef
 import com.badlogic.gdx.physics.box2d.PolygonShape
+import kot.extensions.physics.asBox2DUnits
 
 class FixtureDefBuilder {
